refactor: extract Item component
