+++ wondrous-app/components/Common/Sidebar/Organization/List.tsx
@@ -1,4 +1,4 @@
-import { ButtonBase, Typography } from '@mui/material';
+import { Typography } from '@mui/material';
 import CheckBoxIcon from 'components/Common/Sidebar/Common/icons/checkBox.svg';
 import ContentPaste from 'components/Common/Sidebar/Common/icons/contentPaste.svg';
 import FlagIcon from 'components/Common/Sidebar/Common/icons/flag.svg';
@@ -10,6 +10,7 @@ import PodIcon from 'components/Common/Sidebar/Common/icons/pods.svg';
 import ShowChartIcon from 'components/Common/Sidebar/Common/icons/showChart.svg';
 import StackIcon from 'components/Common/Sidebar/Common/icons/stack.svg';
 import StartIcon from 'components/Common/Sidebar/Common/icons/star.svg';
+import Item from 'components/Common/Sidebar/Common/Item';
 import { useRouter } from 'next/router';
 import styled from 'styled-components';
 import { ENTITIES_TYPES } from 'utils/constants';
@@ -92,81 +93,6 @@ const useDaoSidebarData = () => {
   ];
 };
 
-const ItemButton = styled(ButtonBase)`
-  && {
-    width: 100%;
-    height: 32px;
-    padding: 1px;
-    border-radius: 4px;
-    background: transparent;
-    background: ${({ isActive }) => isActive && 'linear-gradient(90.03deg, #00baff 0.03%, #7427ff 98.82%)'};
-    :hover {
-      background: #313131;
-    }
-  }
-`;
-
-const ItemButtonInner = styled.div`
-  border-radius: 3px;
-  background: transparent;
-  height: 100%;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  gap: 9px;
-  padding: 8px;
-  padding-left: 0px;
-  background: ${({ isActive }) => isActive && '#313131'};
-  ${ItemButton}:hover & {
-    background: #313131;
-  }
-`;
-
-const ItemButtonIcon = styled.div`
-  width: 22px;
-  height: 22px;
-  background: #313131;
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  ${({ isActive }) =>
-    isActive &&
-    `
-      svg {
-     path {
-        stroke: #00baff;
-      }
-      rect {
-        stroke: #00baff;
-      }
-    }
-    `}
-  svg {
-    ${ItemButton}:hover & 
-      path {
-      stroke: #8fe1ff;
-    }
-    ${ItemButton}:hover & 
-      rect {
-      stroke: #8fe1ff;
-    }
-  }
-`;
-
-const ItemButtonText = styled(Typography)`
-  && {
-    font-family: 'Space Grotesk';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 13px;
-    color: ${({ isActive }) => (isActive ? '#00baff' : '#fff')};
-    ${ItemButton}:hover & {
-      color: #8fe1ff;
-    }
-  }
-`;
-
 const Label = styled(Typography)`
   font-family: 'Space Grotesk';
   font-style: normal;
@@ -187,17 +113,6 @@ const ListItemWrapper = styled.div`
   flex-direction: column;
   gap: 36px;
 `;
-
-const Item = ({ children, Icon, isActive, ...props }) => (
-  <ItemButton {...props} disableRipple isActive={isActive}>
-    <ItemButtonInner isActive={isActive}>
-      <ItemButtonIcon isActive={isActive}>
-        <Icon />
-      </ItemButtonIcon>
-      <ItemButtonText isActive={isActive}>{children}</ItemButtonText>
-    </ItemButtonInner>
-  </ItemButton>
-);
 
 const List = () => {
   const daoSidebarData = useDaoSidebarData();
