+++ wondrous-app/components/Common/Sidebar/Settings/styles.tsx
@@ -1,9 +1,5 @@
-import { Box, Button as MuiButton, ButtonBase, InputBase, List, ListItem, Typography } from '@mui/material';
-import SnackbarComp from '@mui/material/Snackbar';
-import { Button } from 'components/Common/button';
+import { Box, List, ListItem, Typography } from '@mui/material';
 import { SafeImage } from 'components/Common/Image';
-import { Discord } from 'components/Icons/discord';
-import { Twitter } from 'components/Icons/twitter';
 import styled from 'styled-components';
 import palette from 'theme/palette';
 
@@ -197,363 +193,4 @@ export const SettingsDaoPodIndicatorIconWrapper = styled.div`
   background: ${({ color }) => color || '#0f0f0f'};
   width: 24px;
   height: 24px;
-`;
-
-// headerBlock.tsx styles
-export const SettingsHeaderBlock = styled.div`
-  position: relative;
-  width: 100%;
-  height: 70px;
-  display: flex;
-  align-items: center;
-`;
-
-export const SettingsHeaderContent = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin-left: 27px;
-`;
-
-export const SettingsHeaderTitle = styled.div`
-  font-weight: 500;
-  font-size: 28px;
-  line-height: 36px;
-  display: flex;
-  align-items: center;
-  color: #ffffff;
-`;
-
-export const SettingsHeaderText = styled(Typography)`
-  && {
-    width: 100%;
-    font-size: 15px;
-    line-height: 23px;
-    letter-spacing: 0.01em;
-    color: #ffffff;
-    display: flex;
-    justify-content: space-between;
-  }
-`;
-
-export const SettingsHeaderInviteButton = styled.button`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  background: #232323;
-  color: #ffffff;
-  font-size: 15px;
-  font-weight: 700;
-  line-height: 150%;
-  padding: 7px 7px 7px 14px;
-  outline: none;
-  border: none;
-  border-radius: 1000px;
-  cursor: pointer;
-`;
-
-export const SettingsHeaderInviteButtonIcon = styled.div`
-  height: 25px;
-  width: 25px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: #3d3d3d;
-  border-radius: 1000px;
-`;
-
-// general settings styles
-export const GeneralSettingsContainer = styled.div`
-  height: 100%;
-  width: 555px;
-`;
-
-export const GeneralSettingsInputsBlock = styled.div`
-  padding: 30px 0;
-  border-bottom: 1px solid #363636;
-`;
-
-export const GeneralSettingsDAONameBlock = styled.div`
-  width: 100%;
-  height: 70px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin-bottom: 28px;
-`;
-export const LabelBlock = styled(Typography)`
-  && {
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 18px;
-    letter-spacing: 0.01em;
-    color: ${palette.blue20};
-    margin-bottom: 10px;
-  }
-`;
-
-export const GeneralSettingsDAONameInput = styled(InputBase)`
-  && {
-    width: 100%;
-    height: 40px;
-    border: 1px solid #4b4b4b;
-    border-radius: 6px;
-
-    font-size: 14px;
-    line-height: 21px;
-    letter-spacing: 0.01em;
-    color: ${palette.white};
-    padding: 10px 15px;
-  }
-`;
-
-export const GeneralSettingsDAODescriptionBlock = styled.div`
-  position: relative;
-`;
-
-export const GeneralSettingsDAODescriptionInputWrapper = styled.div`
-  border: 1px solid pink;
-  border: 1px solid #4b4b4b;
-  border-radius: 6px;
-`;
-
-export const GeneralSettingsDAODescriptionInput = styled(GeneralSettingsDAONameInput)`
-  && {
-    height: 80px;
-    font-size: 14px;
-    line-height: 21px;
-    letter-spacing: 0.01em;
-    color: ${palette.white};
-    padding: 10px 15px;
-    border: none;
-  }
-`;
-
-export const GeneralSettingsDAODescriptionInputCounter = styled(Typography)`
-  && {
-    font-size: 12px;
-    color: #7a7a7a;
-    width: 100%;
-    text-align: right;
-    padding: 8px;
-  }
-`;
-
-// socials block
-export const GeneralSettingsSocialsBlock = styled.div`
-  padding: 30px 0 15px;
-  border-bottom: 1px solid #363636;
-`;
-
-export const GeneralSettingsSocialsBlockWrapper = styled.div``;
-
-export const GeneralSettingsSocialsBlockRow = styled.div`
-  width: 100%;
-  display: flex;
-  margin-bottom: 15px;
-`;
-
-export const GeneralSettingsSocialsBlockRowLabel = styled.div`
-  padding: 7px 12px;
-  background-color: #1c1c1c;
-  border-radius: 4px;
-
-  //text
-  font-size: 14px;
-  line-height: 18px;
-  letter-spacing: 0.01em;
-  text-decoration-line: underline;
-  color: #ccbbff;
-  margin-right: 10px;
-`;
-
-// integrations block
-export const GeneralSettingsIntegrationsBlock = styled.div`
-  width: 100%;
-  padding: 28px 0;
-  border-bottom: 1px solid #363636;
-`;
-
-export const GeneralSettingsDiscordIcon = styled(Discord)`
-  && {
-    width: 27px;
-    height: 20px;
-    color: #00baff;
-    margin-right: 12px;
-  }
-`;
-
-export const GeneralSettingsTwitterIcon = styled(Twitter)`
-  && {
-    width: 27px;
-    height: 20px;
-    color: #00baff;
-    margin-right: 12px;
-  }
-`;
-
-export const GeneralSettingsIntegrationsBlockButton = styled(Button)`
-  && {
-    max-width: 206px;
-  }
-`;
-
-// buttons block
-export const GeneralSettingsButtonsBlock = styled.div`
-  padding: 30px 0;
-  display: flex;
-`;
-
-export const GeneralSettingsResetButton = styled(MuiButton)`
-  && {
-    width: 191px;
-    height: 40px;
-    background: #232323;
-    border-radius: 234px;
-
-    //text
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 150%;
-    color: #ffffff;
-    text-transform: none;
-    font-family: Space Grotesk;
-  }
-`;
-
-export const GeneralSettingsSaveChangesButton = styled(Button)`
-  && {
-    width: 191px;
-    margin-left: 22px;
-    height: 40px;
-    font-family: Space Grotesk;
-    min-height: 40px;
-  }
-`;
-
-// imageUpload.tsx section
-export const ImageUploadBlock = styled.div`
-  width: 100%;
-  padding: 30px 0;
-  border-bottom: 1px solid #363636;
-`;
-
-export const ImageUploadRecommendText = styled(Typography)`
-  && {
-    font-size: 12px;
-    line-height: 15px;
-    color: ${palette.white};
-    margin-bottom: 6px;
-  }
-`;
-
-export const ImageUploadBlockActivitySection = styled.div`
-  display: flex;
-  justify-content: left;
-  flex-direction: column;
-`;
-
-export const ImageUploadBlockInputWrapper = styled.div``;
-
-export const ImageUploadBlockInputButton = styled.input`
-  position: absolute;
-  color: deepskyblue;
-  background-color: transparent;
-  border-bottom: 1px solid deepskyblue;
-  cursor: pointer;
-  display: none;
-`;
-
-export const ImageUploadBlockInputLabel = styled.label`
-  font-size: 14px;
-  line-height: 19px;
-  letter-spacing: 0.01em;
-  text-decoration-line: underline;
-  color: #00baff;
-  cursor: pointer;
-`;
-
-export const ImageUploadBlockUploadedImg = styled.img`
-  margin: 20px 0 10px;
-  border-radius: 4px;
-  object-fit: contain;
-`;
-
-export const ImageUploadBlockRemoveButton = styled(MuiButton)`
-  && {
-    font-size: 12px;
-    line-height: 15px;
-    color: #cb3340;
-    width: 50px;
-    padding: 0 !important;
-    min-width: 0 !important;
-  }
-`;
-
-// LinkSquareIcon.tsx
-export const LinkSquareIconWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: #1c1c1c;
-  border-radius: 3px;
-  margin-right: 30px;
-  padding: 10px;
-`;
-
-// InputField.tsx
-export const LinkInputField = styled(InputBase)`
-  && {
-    width: 333px;
-    height: 32px;
-    border-radius: 6px;
-    border: 1px solid #4b4b4b99;
-    padding-left: 15px;
-
-    //text
-    font-weight: 500;
-    font-size: 15px;
-    line-height: 19px;
-    letter-spacing: 0.01em;
-    text-decoration-line: underline;
-    color: #00baff;
-  }
-`;
-
-export const Snackbar = styled(SnackbarComp)`
-  .MuiPaper-elevation {
-    background: rgb(0, 67, 61);
-    color: white;
-  }
-`;
-
-export const DiscordText = styled(Typography)`
-  && {
-    font-size: 14px;
-    color: ${palette.white};
-    margin-bottom: 8px;
-  }
-`;
-
-export const LabelBlockText = styled(Typography)`
-  && {
-    color: ${palette.white};
-    font-family: Space Grotesk;
-    font-size: 14px;
-    margin-bottom: 12px;
-  }
-`;
-
-export const AddGuildButton = styled(ButtonBase)`
-  && {
-    background: #4b4b51;
-    border-radius: 8px;
-    width: 100%;
-    padding: 8px;
-    height: 40px;
-    display: flex;
-    align-items: center;
-  }
 `;
